+++ imageGenSD.py
@@ -44,7 +44,7 @@ daily_mgr      = None
 daily_mgr_th   = None
 db_ifc         = None
 dict_path      = ["","",""]
-IGSD_version   = '0.3.87'
+IGSD_version   = '0.3.88'
 job_queue      = None
 job_worker     = None
 show_queue     = None
@@ -356,11 +356,15 @@ async def hello(interaction: dis.Interaction):
 @IGSD_client.tree.command()
 @dac.checks.has_permissions(use_application_commands=True)
 @dac.describe(user="The Discord user owning the profiles lsited by the command.  If none, it defaults to you.")
+@dac.describe(name="A name, or fragment of, used to filter the profile list (case insensitive).")
 async def listprofiles(interaction : dis.Interaction,
-                       user        : Optional[dis.User] = None):
+                       user        : Optional[dis.User] = None,
+                       name        : Optional[dac.Range[str, 0, 36]] = None):
     """Returns a pagenated list of profile names and IDs owned by the caller.
 
         Input  : interaction - the interaction context from Discord.
+                 user - optional, the discord user whose profiles to list
+                 name - optional, a name or fragment thereof to filter the profile list
 
         Output : N/A.
     """
@@ -369,13 +373,24 @@ async def listprofiles(interaction : dis.Interaction,
     #This has to be in the function body because an arg can't be used to assign
     #another arg in the function call.
     user_id = interaction.user.id if user == None else user.id
-    profiles = db_ifc.getProfiles(user_id)
+    user_dis = interaction.user if user == None else user
+    error_desc = ""
+    profiles = []
+    message_title = ""
+    if name == None:
+        profiles = db_ifc.getUsersProfiles(user_id)
+        error_desc = f"User <@{user_id}> does not own any characters!"
+        message_title = "Owned characters"
+    else:
+        profiles = db_ifc.getProfiles(user_id, name)
+        error_desc = f"Found no characters owned by user <@{user_id}> that have a name similar to {name}"
+        message_title = f"Owned characters with name like {name}"
     dis_log.debug(f"Got profiles for list profile: {profiles}.")
 
     if not profiles:
-
-        embed = dis.Embed(title       = "Owned characters",
-                          description = f"User <@{user_id}> does not own any characters!",
+        
+        embed = dis.Embed(title       = message_title,
+                          description = error_desc,
                           color       = 0xec1802)
         await interaction.response.send_message(content=f"<@{interaction.user.id}>", embed=embed)
 
@@ -385,7 +400,7 @@ async def listprofiles(interaction : dis.Interaction,
 
         await mp.MenuPagination(interaction = interaction,
                                 profiles    = short_profiles,
-                                user        = user).navigate()
+                                user        = user_dis).navigate()
 
 @IGSD_client.event
 async def on_ready():
@@ -597,16 +612,19 @@ async def roll(interaction: dis.Interaction):
 
 @IGSD_client.tree.command()
 @dac.checks.has_permissions(use_application_commands=True)
-@dac.describe(profile_id="The profile ID of the character you'd like to view.  Use /listprofiles to see the name and ID for profiles!")
-@dac.describe(user="The Discord user owning the profiles listed by the command.  If none, it defaults to you.")
+@dac.describe(profile_id="The profile ID of the character.  Use /listprofiles to find the ID.")
+@dac.describe(user="The Discord user owning the profiles.  If none, it defaults to you.")
+@dac.describe(name="A name of the profile to view. Will find all similar names (case insensitive).")
 async def showprofile(interaction : dis.Interaction,
                       profile_id  : Optional[dac.Range[str, 0, 36]] = None,  #The length of a UUID
-                      user        : Optional[dis.User] = None):
-    """Displays the profile associated with the provided ID.
+                      user        : Optional[dis.User] = None,
+                      name        : Optional[dac.Range[str, 0, 36]] = None):
+    """Displays a profile or a dropdown of matching profiles.
 
         Input  : interaction - the interaction context from Discord.
                  profile_id - the profile ID to retrieve.
                  user - which user's profiles to show; defaults to the author.
+                 name - A name that will be used to search the profiles
 
         Output : N/A.
     """
@@ -620,27 +638,10 @@ async def showprofile(interaction : dis.Interaction,
                }
 
 
+    user_id  = interaction.user.id if user == None else user.id
 
-    if profile_id == None:
-
-        if db_ifc.getDropdown(user_id = interaction.user.id) :
-
-            await interaction.response.send_message(f'Please close your existing dropdown menu or wait for it to time out.', ephemeral=True, delete_after=9.0)
-
-        else:
-
-            dis_log.debug(f"Creating a SHOW PROFILE view for user {interaction.user.id}.")
-            user_id  = interaction.user.id if user == None else user.id
-            profiles = db_ifc.getProfiles(user_id)
-            dis_log.debug(f"User {user_id} returned profile list {profiles}.")
-            view = ddf.DropdownView(ctx      = interaction,
-                                    type     = ddf.DropDownTypeEnum.SHOW,
-                                    choices  = profiles,
-                                    metadata = metadata)
-
-            await interaction.response.send_message('Select a profile to view:',view=view)
-
-    else:
+    if profile_id != None:
+        #The profile id was provided, prioritize that over everything else
 
         opts = {'id' : profile_id}
 
@@ -652,8 +653,65 @@ async def showprofile(interaction : dis.Interaction,
         result = show_queue.add(metadata = metadata,
                                 job      = job)
 
-        await interaction.response.send_message(f'{result}', ephemeral=True, delete_after=9.0)
+        await interaction.response.send_message(f'{result}')
+    else:
+        #get profiles from the database using user and name
 
+        if db_ifc.getDropdown(user_id = interaction.user.id) :
+
+            await interaction.response.send_message(f'Please close your existing dropdown menu or wait for it to time out.', ephemeral=True, delete_after=9.0)
+
+        else:
+
+            dis_log.debug(f"Creating a SHOW PROFILE view for user {interaction.user.id}.")
+            empty_error = "";
+            many_message = "";
+
+            if name == None:
+
+                profiles = db_ifc.getUsersProfiles(user_id)
+                empty_error = f"<@{user_id}> does't own any profiles"
+                many_message = f'Select a profile to view:'
+
+            else:
+
+                dis_log.debug(f"Looking for profiles of name {name}.")
+                profiles = db_ifc.getProfiles(user_id, name)
+                empty_error = f'Found no characters owned by <@{user_id}> that have a name similar to {name}'
+                many_message = f'Select a profile with name like "{name}" to view:'
+
+            if len(profiles) == 0:
+
+                dis_log.debug(f"Looking for name, Found none {name}")
+                embed = dis.Embed(title="Error",
+                                  description=empty_error,
+                                  color=0xec1802)
+                await interaction.response.send_message(content=f"<@{interaction.user.id}>", embed=embed)
+
+            elif len(profiles) == 1:
+
+                dis_log.debug(f"Looking for name, Found one {name} {profiles[0].id}")
+                opts = {'id' : profiles[0].id}
+
+                dis_log.debug(f"Creating a job with metadata {metadata} and options {opts}.")
+                job = jf.JobFactory.getJob(type    = jf.JobTypeEnum.SHOW_PROFILE,
+                                           ctx     = interaction,
+                                           options = opts)
+                dis_log.debug(f"Posting SHOW job {job} to the queue.")
+                result = show_queue.add(metadata = metadata,
+                                        job      = job)
+
+                await interaction.response.send_message(f'{result}')
+
+            else:
+
+                dis_log.debug(f"Looking for name, Found many {len(profiles)}")
+                view = ddf.DropdownView(ctx      = interaction,
+                                        type     = ddf.DropDownTypeEnum.SHOW,
+                                        choices  = profiles,
+                                        metadata = metadata)
+
+                await interaction.response.send_message(many_message,view=view)
 
 @verify(UNIQUE)
 class SummaryChoices(Enum):