+++ src/db/MariadbIfc.py
@@ -317,7 +317,6 @@ class MariadbIfc:
 
         return result
 
-
     def getImage(self,
                  picture_id : Optional[str] = None,
                  profile_id : Optional[str] = "ffffffff-ffff-ffff-ffff-fffffffffffe") -> str:
@@ -401,20 +400,22 @@ class MariadbIfc:
         return profile
 
     def getProfiles(self,
-                    user_id : int) -> list:
-        """Returns all profiles for a given user.
+                    user_id : int,
+                    name : str) -> list:
+        """Returns profiles matching a given name filter for a given user.
 
             Input: self - Pointer to the current object instance.
                    user_id - user ID to interrogate for profiles.
+                   name - string to match like the profile name.
 
-            Output: list - A list of all profiles found, if any.  None if not.
+            Output: list - A list of all profiles found, if any.  An empty list if not.
         """
         cmd     = ""
         cursor  = self.con.cursor(buffered=False)
         results = []
 
-        self.db_log.info(f"Getting profiles for user {user_id}")
-        cmd = (self.cmds['prof']['get_owned_profs']) % (user_id)
+        self.db_log.info(f"Getting profiles matching {name} for user {user_id}")
+        cmd = (self.cmds['prof']['get_owned_profs_byname']) % (user_id, f"%{name}%")
         self.db_log.debug(f"Executing command: {cmd}")
         cursor.execute(cmd)
 
@@ -699,6 +700,33 @@ class MariadbIfc:
 
         self.db_log.info(f"Getting unoccupied profiles for user {user_id}")
         cmd = (self.cmds['prof']['get_unoccupied_profs']) % (user_id)
+        self.db_log.debug(f"Executing command: {cmd}")
+        cursor.execute(cmd)
+
+        for x in cursor:
+
+            self.db_log.debug(f"Adding result: {x}")
+            results.append(self.mapQueryToProfile(query=x))
+
+        self.db_log.debug(f"Got results: {results}")
+
+        return results
+
+    def getUsersProfiles(self,
+                         user_id : int) -> list:
+        """Returns all profiles for a given user.
+
+            Input: self - Pointer to the current object instance.
+                   user_id - user ID to interrogate for profiles.
+
+            Output: list - A list of all profiles found, if any.  An empty list if not.
+        """
+        cmd     = ""
+        cursor  = self.con.cursor(buffered=False)
+        results = []
+
+        self.db_log.info(f"Getting profiles for user {user_id}")
+        cmd = (self.cmds['prof']['get_owned_profs']) % (user_id)
         self.db_log.debug(f"Executing command: {cmd}")
         cursor.execute(cmd)